[refactor]: seperated exchangeItem widget and fixed bottom sheet animation
